Extract related logic into OOP classes
