+++ src/main/java/diheng/DiHengException.java
@@ -1,3 +1,4 @@
+package diheng;
 
 public class DiHengException extends Exception {
