+++ src/main/java/diheng/enums/Command.java
@@ -1,3 +1,4 @@
+package diheng.enums;
 public enum Command {
     LIST,
     TODO,
