+++ src/main/java/diheng/tasks/Event.java
@@ -1,4 +1,4 @@
-
+package diheng.tasks;
 public class Event extends Task {
 
     private final String start;
