+++ src/main/java/diheng/tasks/Task.java
@@ -1,4 +1,4 @@
-
+package diheng.tasks;
 public class Task {
 
     private final String description;
